Add Logo in docs pages
